hwc: set display active flag in its own blank
For non-primary display, if driver reports valid display attribute
in the boot up, then SF will treat it as unblanked already, hence
no unblank will be called for this display. So we need to set it
as active when primary display is unblanked. But subsequent unblank
call is made for each display during suspend/resume, hence we need
to set active flag per display.

Change-Id: I8ff175eca326044e3afd946918cc32f5ff7e12c7